+++ test/functional/end.spec.ts
@@ -1,7 +1,7 @@
 import chai, { assert } from "chai";
 import chaiHttp from "chai-http";
 
-import app from "../../index";
+import app from "../../src/index";
 
 chai.use(chaiHttp);
