+++ test/functional/index.spec.ts
@@ -1,8 +1,8 @@
 import chai, { assert } from "chai";
 import chaiHttp from "chai-http";
 
-import app from "../../index";
-import { SnakeInfoInt } from "../../interfaces/SnakeInfoInt";
+import app from "../../src/index";
+import { SnakeInfoInt } from "../../src/interfaces/SnakeInfoInt";
 
 chai.use(chaiHttp);
