+++ test/unit/findObstacleTest.spec.ts
@@ -1,9 +1,9 @@
 import { assert } from "chai";
 
-import { BattleSnakeInt } from "../../interfaces/BattleSnakeInt";
-import { BoardInt } from "../../interfaces/BoardInt";
-import { RequestBodyInt } from "../../interfaces/RequestBodyInt";
-import { findOccupiedSquares } from "../../modules/findOccupiedSquares";
+import { BattleSnakeInt } from "../../src/interfaces/BattleSnakeInt";
+import { BoardInt } from "../../src/interfaces/BoardInt";
+import { RequestBodyInt } from "../../src/interfaces/RequestBodyInt";
+import { findOccupiedSquares } from "../../src/modules/findOccupiedSquares";
 
 const mockSnake: BattleSnakeInt = {
   id: "mockEnemy",
